Recoded import of state transition diagrams in XML format
